backend/requirements.txt
azure-functions
azure-storage-blob
pandas
scikit-learn
azure-cosmos
